requirements.txt
telethon
python-dotenv
psycopg2-binary
apscheduler
telegraph
